dev: split def/expr, refactor definition
