+++ crates/tinymist-derive/src/lib.rs
@@ -46,3 +46,59 @@ pub fn bind_ty_ctx(input: TokenStream) -> TokenStream {
     // Hand the output tokens back to the compiler
     TokenStream::from(expanded)
 }
+
+#[proc_macro_derive(DeclEnum)]
+pub fn gen_decl_enum(input: TokenStream) -> TokenStream {
+    // In form of
+    // ```
+    // pub enum Decl {
+    //   Sub1(X),
+    //   Sub2(Y),
+    // }
+    // ```
+
+    // Parse the input tokens into a list of variants
+    let input = parse_macro_input!(input as DeriveInput);
+
+    let variants = match input.data {
+        syn::Data::Enum(data) => data.variants,
+        _ => panic!("only enums are supported"),
+    };
+
+    let names = variants.iter().map(|v| &v.ident).collect::<Vec<_>>();
+
+    let input_name = &input.ident;
+
+    let expanded = quote! {
+        impl #input_name {
+            pub fn name(&self) -> &Interned<str> {
+                match self {
+                    #(Self::#names(x) => x.name()),*
+                }
+            }
+
+            pub fn span(&self) -> Option<Span> {
+                match self {
+                    #(Self::#names(x) => x.span()),*
+                }
+            }
+
+            pub fn kind(&self) -> DefKind {
+                match self {
+                    #(Self::#names(..) => DefKind::#names),*
+                }
+            }
+        }
+
+        impl fmt::Debug for Decl {
+            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+                match self {
+                    #(Self::#names(x) => write!(f, concat!(stringify!(#names), "({:?})"), x)),*
+                }
+            }
+        }
+
+    };
+
+    TokenStream::from(expanded)
+}
